added wez's "expression" code which extends syntax to MIDL-compliant: [size_is(something * something_else - yet more)]
fixed issues with flex 2.5.x which are apparently also fixed in
the dce-rpc 1 release made by loic last month, in a different
way.  what i added is an unexplainable hack that happens to work.

added "imports" which supports multiple interfaces: not entirely
sure who to attribute this to.

TODO: double-check the yacc dceidl error messages.