MSRPC limits size of switch to 2 bytes - even for longs.
result: pain in the arse.  MSRPC leaves the 2 bytes after the
switch value uninitialised, such that any non-encapsulated
union structure with a size of four or more bytes leaves
DCE/RPC thinking that a long switch value is gobbledegook in the
top two bytes.

this is alleviated by a) &= 0xffff'ing the switch value

b) writing _back_ the truncated switch value into the data being
unmarshalled.  this is important: it's no good unpacking the structures
correctly only to leave the case statement switch value of the
non-encapsulated union as complete bollocks.