this may look like a big commit but it's not: the ee*.c files are near-identical.
i've added (and disabled) some code which turns "full" pointers
into "unique" ones - in the most awful hacked-up way i could
possibly get away with.

the data stream being unmarshalled is MODIFIED when a pointer token
is discovered, to be replaced with a unique number.

in order to avoid messing with the code too much, i reserve a range
of pointer tokens up to 65536.

it's all very horrible but it's the only way with a minimal amount
of code changing to obtain pointer-interoperability with MSRPC,
especially as MS successfully added non-encapsulated union
unique pointers [to MIDL].

the way they would have done it is to either follow the
[pointer] codepath even for unique pointers, which has a
performance penalty (it requires an extra pass over the
data stream to fill in the pointees), or to waste memory by
allocating the maximum size of the union.

don't know, don't care - this approach works for me, and it's
also paves the way for doing a proper job.