oh _man_ this one's a whole boat-load of trouble.
* renamed all pthread_xxx functions to sys_pthread_xxx with a little
  automated script

* created a #define pthreads_rename.h which defines all functions
  sys_pthread_xx to pthd4_xx

* updated all (original) uses of pthread_xxx to conform to posix
  STANDARD pthreads NOT posix draft 4 pthreads.

so it's a bit weird, and yes it involved a lot of testing on the way,
and yes each stage worked.

i also updated dcethreads - removing the use of a Version file which
redirects pthread_xxx to pthd4_xxx - so you _have_ to explicitly use
the functions pthd4_xxx you can't get away with just trying to use
pthread_xxx "looking" like the same named posix thread function.
you _have_ to use the pthd4_xxx function which is what the
pthreads_rename.h is all about.

... with me so far? :)

* removed all "direct" references to pthread_setcancelstate and
  pthread_setcanceltype and utilised the pthread_setcancel and
  pthread_setasynccancel functions provided in libdcethreads
  instead.

* oh - and i removed the references to #ifdef ENABLE_PTHREADS because
  they're no longer necessary: because the sys_pthread_xxx functions
  all now conform to the posix thread standard (not draft 4).