add in the start of a socket abstraction layer, such that it will be possible to emulate sockets behind the rpc__socket_*() functions.
the aim is to make it easy to add NamedPipes behind sockets, and also
oh i dunno, an implementation of ncalrpc over shared memory, all that
sort of good stuff.

it's therefore necessary to rip out some stuff like the use of
getsockname(), getpeername() etc. and the start of that is the
creation of an rpc__socket_nodelay() function, which is in, for
some weird reason, ipnaf.c.  huhn.